# docs/openapi/permesi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "permesi",
    "description": "Identity and Access Management",
    "contact": {
      "name": "Team Permesi",
      "email": "team@permesi.dev"
    },
    "license": {
      "name": "BSD-3-Clause",
      "identifier": "BSD-3-Clause"
    },
    "version": "0.3.43"
  },
  "paths": {
...
    "/v1/auth/opaque/login/finish": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "opaque_login_finish",
        "parameters": [
          {
            "name": "X-Permesi-Zero-Token",
            "in": "header",
            "description": "Genesis zero token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpaqueLoginFinishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Login success"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/opaque/login/start": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "opaque_login_start",
        "parameters": [
          {
            "name": "X-Permesi-Zero-Token",
            "in": "header",
            "description": "Genesis zero token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpaqueLoginStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OPAQUE login started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpaqueLoginStartResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/auth/session": {
      "get": {
        "tags": [
          "auth"
        ],
        "operationId": "session",
        "responses": {
          "200": {
            "description": "Session is active",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "204": {
            "description": "No active session"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OpaqueLoginFinishRequest": {
        "type": "object",
        "required": [
          "login_id",
          "email",
          "credential_finalization"
        ],
        "properties": {
          "credential_finalization": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "login_id": {
            "type": "string"
          }
        }
      },
      "OpaqueLoginStartRequest": {
        "type": "object",
        "required": [
          "email",
          "credential_request"
        ],
        "properties": {
          "credential_request": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "OpaqueLoginStartResponse": {
        "type": "object",
        "required": [
          "login_id",
          "credential_response"
        ],
        "properties": {
          "credential_response": {
            "type": "string"
          },
          "login_id": {
            "type": "string"
          }
        }
      },
...
      "SessionKind": {
        "type": "string",
        "description": "Session kinds used to gate MFA bootstrap and challenge flows.",
        "enum": [
          "full",
          "mfa_bootstrap",
          "mfa_challenge"
        ]
      },
      "SessionResponse": {
        "type": "object",
        "required": [
          "user_id",
          "email",
          "is_operator",
          "session_kind",
          "totp_enabled",
          "webauthn_enabled"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "is_operator": {
            "type": "boolean"
          },
          "session_kind": {
            "$ref": "#/components/schemas/SessionKind"
          },
          "totp_enabled": {
            "type": "boolean"
          },
          "user_id": {
            "type": "string"
          },
          "webauthn_enabled": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}